+++ src/Command/CommandInterface.php
@@ -829,18 +829,26 @@ interface CommandInterface
      *
      * @param string $table The name of the table to update.
      * @param array $columns The column data (name => value) to update.
-     * @param array|string $condition The condition to put in the WHERE part. Please refer to
-     * {@see QueryInterface::where()} on how to specify condition.
+     * @param array|ExpressionInterface|string $condition The condition to put in the WHERE part. Please refer to
+     * {@see QueryPartsInterface::where()} on how to specify condition.
+     * @param array|ExpressionInterface|string|null $from The FROM part. Please refer to {@see QueryPartsInterface::from()}
+     * on how to specify FROM part.
      * @param array $params The parameters to bind to the command.
+     *
+     * @psalm-param ParamsType $params
      *
      * @throws Exception
      * @throws InvalidArgumentException
      *
-     * @psalm-param ParamsType $params
-     *
      * Note: The method will quote the `table` and `columns` parameter before using it in the generated SQL.
      */
-    public function update(string $table, array $columns, array|string $condition = '', array $params = []): static;
+    public function update(
+        string $table,
+        array $columns,
+        array|ExpressionInterface|string $condition = '',
+        array|ExpressionInterface|string|null $from = null,
+        array $params = []
+    ): static;
 
     /**
      * Creates a command to insert rows into a database table if they don't already exist (matching unique constraints)