+++ src/QueryBuilder/DMLQueryBuilderInterface.php
@@ -10,6 +10,7 @@ use Yiisoft\Db\Exception\Exception;
 use InvalidArgumentException;
 use Yiisoft\Db\Exception\InvalidConfigException;
 use Yiisoft\Db\Exception\NotSupportedException;
+use Yiisoft\Db\Expression\ExpressionInterface;
 use Yiisoft\Db\Query\QueryInterface;
 
 /**
@@ -177,8 +178,10 @@ interface DMLQueryBuilderInterface
      *
      * @param string $table The table to update.
      * @param array $columns The column data (name => value) to update the table.
-     * @param array|string $condition The condition to put in the `WHERE` part. Please refer to
-     * {@see Query::where()} On how to specify condition.
+     * @param array|ExpressionInterface|string $condition The condition to put in the `WHERE` part. Please refer to
+     * {@see QueryPartsInterface::where()} On how to specify condition.
+     * @param array|ExpressionInterface|string|null $from The FROM part. Please refer to
+     * {@see QueryPartsInterface::from()} On how to specify FROM part.
      * @param array $params The binding parameters that will be modified by this method so that they can be bound to
      * DB command later.
      *
@@ -191,7 +194,13 @@ interface DMLQueryBuilderInterface
      *
      * Note: The method will escape the table and column names.
      */
-    public function update(string $table, array $columns, array|string $condition, array &$params = []): string;
+    public function update(
+        string $table,
+        array $columns,
+        array|ExpressionInterface|string $condition,
+        array|ExpressionInterface|string|null $from = null,
+        array &$params = []
+    ): string;
 
     /**
      * Creates an SQL statement to insert rows into a database table if they don't already exist (matching unique