+++ tests/Db/QueryBuilder/QueryBuilderTest.php
@@ -243,7 +243,8 @@ final class QueryBuilderTest extends AbstractQueryBuilderTest
     public function testUpdate(
         string $table,
         array $columns,
-        array|string $condition,
+        array|ExpressionInterface|string $condition,
+        array|ExpressionInterface|string|null $from,
         array $params,
         string $expectedSql,
         array $expectedParams = [],
@@ -251,11 +252,13 @@ final class QueryBuilderTest extends AbstractQueryBuilderTest
         $db = $this->getConnection();
         $qb = $db->getQueryBuilder();
 
-        $sql = $qb->update($table, $columns, $condition, $params);
+        $sql = $qb->update($table, $columns, $condition, $from, $params);
         $sql = $db->getQuoter()->quoteSql($sql);
 
         $this->assertSame($expectedSql, $sql);
         $this->assertEquals($expectedParams, $params);
+
+        $db->close();
     }
 
     #[DataProviderExternal(QueryBuilderProvider::class, 'upsert')]